pyproject.toml
[project]
name = "yt-cut"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "aiogram",
    "python-dotenv",
]
